peds_unity_system: peds run away from side when vehicle comes; fixed pedestrian tag bug; fixed vehicle vel bug

+++ peds_unity_system/src/RVO2/src/Agent.h
@@ -112,7 +112,8 @@ namespace RVO {
 		int change_dir_iter_;
 		Vector2 new_pref_vel_;
 
-		size_t ped_id_; // this is to track whether this pedestrian, for external use, different from id_;
+		//size_t ped_id_; // this is to track whether this pedestrian, for external use, different from id_;
+		int ped_id_;
 		bool updated_;
 	};
